(refactor) extract ReviewPage functionality to useReviewPage component hook
